feat(graphin):move Behaviors into graphin
